+++ src/main/java/org/thingai/app/scoringservice/callback/EventHandlerCallback.java
@@ -1,6 +1,6 @@
 package org.thingai.app.scoringservice.callback;
 
-import org.thingai.app.scoringservice.entity.event.Event;
+import org.thingai.app.scoringservice.entity.Event;
 
 public interface EventHandlerCallback {
     void onSetEvent();
