+++ src/main/java/org/thingai/app/scoringservice/dto/MatchDetailDto.java
@@ -1,9 +1,9 @@
 package org.thingai.app.scoringservice.dto;
 
-import org.thingai.app.scoringservice.entity.match.AllianceTeam;
-import org.thingai.app.scoringservice.entity.match.Match;
-import org.thingai.app.scoringservice.entity.score.Score;
-import org.thingai.app.scoringservice.entity.team.Team;
+import org.thingai.app.scoringservice.entity.AllianceTeam;
+import org.thingai.app.scoringservice.entity.Match;
+import org.thingai.app.scoringservice.entity.Score;
+import org.thingai.app.scoringservice.entity.Team;
 
 public class MatchDetailDto {
     private Match match;
