+++ src/main/java/org/thingai/app/scoringservice/entity/RankingSnapshot.java
@@ -1,4 +1,4 @@
-package org.thingai.app.scoringservice.entity.ranking;
+package org.thingai.app.scoringservice.entity;
 
 public class RankingSnapshot {
     private int matchType;
