+++ src/main/java/org/thingai/app/scoringservice/repository/DaoAllianceTeam.java
@@ -1,6 +1,6 @@
 package org.thingai.app.scoringservice.repository;
 
-import org.thingai.app.scoringservice.entity.match.AllianceTeam;
+import org.thingai.app.scoringservice.entity.AllianceTeam;
 import org.thingai.base.dao.Dao;
 
 public class DaoAllianceTeam {
